requirements.txt
python-dotenv
openai
faiss-cpu
langchain
pydantic
langchain-community